Extract `fetchMessages`
